distutils: Support FOO_dll.a libraries in EMX toolchain.
This is needed to e.g. support new sqlite RPMs where the import library is named sqlite3_dll.a (w/o the lib prefix).

Note also that the rest of the EMX toolchain (e.g. emxomfld) supports that as well. Ideally, distutils should be always kept in sync with it.